% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CornishFisher.Method.R
\name{CornishFisher.Method}
\alias{CornishFisher.Method}
\title{Univariate Cornish Fisher Method}
\usage{
CornishFisher.Method(data, confidence.level = 0.95)
}
\arguments{
\item{data}{Daily Log-Return Series}

\item{confidence.level}{Confidence Level for Risk calculation,e.g. 95\%}
}
\value{
Value-at-Risk and Expected Shortfall Estimates
}
\description{
It performs Cornish Fisher method for univariate daily risk calculation.
}
\examples{
CornishFisher.Method(rnorm(100,2,1),confidence.level=0.95)

}
