% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exceptions.R
\name{Exceptions}
\alias{Exceptions}
\title{Exceptions for Back-Testing}
\usage{
Exceptions(VaR, data)
}
\arguments{
\item{VaR}{Value-at-Risk estimates for n days}

\item{data}{Daily Log Returns Data of the n days}
}
\value{
A vector with 0's for the non-exceptions days and daily returns in exceptions days
}
\description{
Generates a vector with 0's in the days in which exceptions do not occur
and the daily return in the ones where the daily return exceeds VaR estimates
}
\examples{
Exceptions(VaR.data,actual.data)

}
