% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.Parameters.R
\name{Fit.Parameters}
\alias{Fit.Parameters}
\title{Fit Parameters}
\usage{
Fit.Parameters(data, dist = c("normal", "std", "skewn", "skewt"))
}
\arguments{
\item{dist}{"normal","std","skewn","skewt"}
}
\value{
Parameters of the fitted distributions based on the choice
}
\description{
For the input data, it finds the parameters for "Standard Normal", "Standard-t", "Skewed-Normal" and "Skewed-t" distributions
}
\examples{
Fit.Parameters(rnorm(100,1,2),dist="normal")

}
