% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Garch.Simulation.R
\name{Garch.Simulation}
\alias{Garch.Simulation}
\title{GARCH Data Generation Process}
\usage{
Garch.Simulation(data, spec, n.sim = 1000, horizon = 1)
}
\arguments{
\item{data}{Log-return series}

\item{spec}{Specification of GARCH model  for the log-return series with ARMA(p,q) and GARCH(m,n) orders}

\item{n.sim}{Number of Simulations}

\item{horizon}{Daily Horizon, 1 for 1-day, 10 for 10-day}
}
\value{
Simulated Forecasts
}
\description{
When the input is the log-return series, it simulates new time series by using the input specification of ARMA(p,q)-GARCH(p,q) process for the pre-specified time horizon
}
\examples{
garchspec <- ugarchspec(variance.model = list(model="sGARCH",garchOrder=c(1,1)),
mean.model = list(armaOrder=c(0,0),include.mean=TRUE))
Garch.Simulation(rnorm(100,1,2),garchspec,n.sim=1000,horizon=1)
}
