% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Historical.Simulation.R
\name{Historical.Simulation}
\alias{Historical.Simulation}
\title{Historical Simulation Method}
\usage{
Historical.Simulation(data, confidence.level = 0.95)
}
\arguments{
\item{data}{Vector of Asset Returns}

\item{confidence.level}{confidence level for the risk calculation, e.g. 95\%}
}
\value{
VaR and ES estimates based on a risk level
}
\description{
Historical Simulation Method
}
\examples{
Historical.Simulation(rnorm(100,-0.5,1),confidence.level=0.95)
}
