% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.Method.R
\name{MonteCarlo.Method}
\alias{MonteCarlo.Method}
\title{Univariate Monte-Carlo Simulation Method}
\usage{
MonteCarlo.Method(
  data,
  dist = c("normal", "std", "skewn", "skewt"),
  n.sim = 1000,
  confidence.level = 0.95
)
}
\arguments{
\item{data}{Daily Log-Return Series}

\item{dist}{"normal","std","skewn","skewt"}

\item{n.sim}{Number of Simulations}

\item{confidence.level}{Confidence Level for Risk calculation,e.g. 95\%}
}
\value{
Value-at-Risk and Expected Shortfall Estimates
}
\description{
It performs Monte-Carlo method for univariate daily risk calculation based on the selected distribution.
}
\examples{
MonteCarlo.Method(rnorm(100,1,2),dist="normal",n.sim=1000,confidence.level=0.95)

}
