% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametric.Method.R
\name{Parametric.Method}
\alias{Parametric.Method}
\title{Univariate Parametric Risk Method}
\usage{
Parametric.Method(
  data,
  dist = c("normal", "std", "skewn", "skewt"),
  confidence.level = 0.95
)
}
\arguments{
\item{data}{Daily Log-Return Series}

\item{dist}{"normal","std","skewn","skewt"}

\item{confidence.level}{Confidence Level for Risk calculation,e.g. 95\%}
}
\value{
Value-at-Risk and Expected Shortfall Estimates
}
\description{
It performs Parametric method for univariate daily risk calculation based on the selected distribution.
}
