% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Distributions.R
\name{Plot.Distributions}
\alias{Plot.Distributions}
\title{Density Plots on Histogram}
\usage{
Plot.Distributions(data)
}
\arguments{
\item{data}{Return Data}
}
\value{
Plots the density functions on Histogram
}
\description{
Data is fitted into the Standard Normal, Standard-t, Skewed Normal and Skewed t and then they are plotted on the histogram with different colors.
}
\examples{
Plot.Distributions(rnorm(100,1,2))
}
