% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price.Data.R
\name{Price.Data}
\alias{Price.Data}
\title{Daily Price Data Loading}
\usage{
Price.Data(
  start = "2023-01-01",
  end = "2024-01-01",
  tickers = c("AAPL", "META"),
  type = c("Open", "High", "Low", "Close", "Volume", "Adjusted")
)
}
\arguments{
\item{start}{First date of the data}

\item{end}{Last date of the data}

\item{tickers}{Asset Tickers}

\item{type}{Open,High,Low,Close,Volume,Adjusted}
}
\value{
Daily Prices
}
\description{
Daily Price Data Loading
}
\examples{
Price.Data(start="2023-01-01",end="2024-01-01",tickers="AAPL",type="Close")

}
