% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RichiCeretta.Backtesting.R
\name{RichiCeretta.Backtesting}
\alias{RichiCeretta.Backtesting}
\title{Righi and Ceretta Expected Shortfall Backtesting}
\usage{
RichiCeretta.Backtesting(
  risk.var,
  data,
  confidence.level = 0.95,
  n.boots,
  n.sim = 10000,
  garch.spec
)
}
\arguments{
\item{risk.var}{Value-at-Risk Estimates for the out-of-sample}

\item{data}{Return data together with in-sample and out-of-sample}

\item{confidence.level}{Confidence Level for Back-Testing,e.g. 95\%}

\item{n.boots}{Size of the Bootstrapping Samples}

\item{n.sim}{Number of Simulations (Bootstrap Replicates)}

\item{garch.spec}{Specification of GARCH process that is required for GARCH simulation}
}
\value{
The test statistics and p-values for the days where the exceptions occur
}
\description{
Performs Righi and Ceretta (2013) Expected Shortfall Backtesting
}
\examples{
garchspec <- ugarchspec(variance.model = list(model="sGARCH",garchOrder=c(1,1)),
mean.model = list(armaOrder=c(0,0),include.mean=TRUE))

getSymbols("^HSI",from="2017-06-07",to="2018-09-02",src="yahoo")
returns <- as.vector(na.omit(diff(log(Ad(`HSI`)))))
hsforecasts <- Rolling.Window(data=returns,window.size = 250,func = Historical.Simulation,confidence.level = 0.95)
hsforecasts <- do.call(rbind,hsforecasts)
varforecasts <- hsforecasts$Value_at_Risk
esforecasts <- hsforecasts$Expected_Shortfall
RichiCeretta.Backtesting(risk=varforecasts,data=returns,garch.spec=garchspec,n.boots=10000,n.sim=1000)

}
