% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR.Backtesting.R
\name{VaR.Backtesting}
\alias{VaR.Backtesting}
\title{Value-at-Risk Backtesting Conditional Coverage Test, Independence Test and Kupiec Test}
\usage{
VaR.Backtesting(excepts, confidence.level = 0.95)
}
\arguments{
\item{excepts}{Exceptions data calculated by Exceptions Function}

\item{confidence.level}{Confidence Level for Back-Testing,e.g. 95\%}
}
\value{
Test Statistics, Critical Values, p-values, Zero Hypothesis and the Test Decisions
}
\description{
It performs Kupiec's Exceedances Test, Independence Test and Conditional Coverage (Joint)
Test for Value-at-Risk Backtesting
}
\examples{
VaR.Backtesting(exceptions.data,confidence.level=0.95)

}
